mtaka-backend/requirements.txt
django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
mysqlclient==2.2.8
psycopg2-binary==2.9.10
Pillow==11.3.0
gunicorn==23.0.0
whitenoise==6.11.0
